relocate all wordnik property classes to their own WordnikProperties file
