+++ src/main/kotlin/systems/ajax/englishstudytelegrambot/config/Configuration.kt
@@ -1,13 +1,12 @@
 package systems.ajax.englishstudytelegrambot.config
 
-import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.boot.context.properties.bind.ConstructorBinding
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.reactive.function.client.WebClient
-import systems.ajax.englishstudytelegrambot.service.WordinkKeyProperty
-import systems.ajax.englishstudytelegrambot.service.WordnikLinkProperties
+import systems.ajax.englishstudytelegrambot.property.WordinkKeyProperty
+import systems.ajax.englishstudytelegrambot.property.WordnikBaseUrlProperty
+import systems.ajax.englishstudytelegrambot.property.WordnikLinkProperties
 
 @Configuration
 @EnableConfigurationProperties(WordnikLinkProperties::class, WordnikBaseUrlProperty::class, WordinkKeyProperty::class)
@@ -18,8 +17,3 @@ class Configuration {
             .baseUrl(wordnikBaseUrlProperty.url)
             .build()
 }
-
-@ConfigurationProperties(prefix = "base")
-data class WordnikBaseUrlProperty @ConstructorBinding constructor(
-    val url: String
-)
