+++ src/main/kotlin/systems/ajax/englishstudytelegrambot/service/WordServices.kt
@@ -1,32 +1,20 @@
 package systems.ajax.englishstudytelegrambot.service
 
 import org.slf4j.LoggerFactory
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.bind.ConstructorBinding
-import org.springframework.http.HttpStatus
-import org.springframework.stereotype.Component
 import org.springframework.stereotype.Service
-import org.springframework.web.reactive.function.client.WebClient
 import org.springframework.web.server.ResponseStatusException
 import systems.ajax.englishstudytelegrambot.dto.AudioForWordResponse
 import systems.ajax.englishstudytelegrambot.dto.DefinitionOfWordResponse
 import systems.ajax.englishstudytelegrambot.dto.ExampleOfWordResponse
 import systems.ajax.englishstudytelegrambot.dto.PronunciationOfWordResponse
-import systems.ajax.englishstudytelegrambot.dto.GettingPartOfAdditionalInfoAboutWord
 import systems.ajax.englishstudytelegrambot.entity.AdditionalInfoAboutWord
+import systems.ajax.englishstudytelegrambot.external.source.ExternalWordSource
+import systems.ajax.englishstudytelegrambot.property.WordnikLinkProperties
 import kotlin.reflect.KSuspendFunction1
 
 interface AdditionalInfoAboutWordService {
     suspend fun findAdditionInfoAboutWord(wordSpelling: String): AdditionalInfoAboutWord
 }
-
-@ConfigurationProperties(prefix = "link")
-data class WordnikLinkProperties @ConstructorBinding constructor(
-    val audio: String,
-    val definition: String,
-    val example: String,
-    val pronunciation: String
-)
 
 @Service
 class AdditionalInfoAboutWordServiceImpl(
@@ -75,27 +63,4 @@ class AdditionalInfoAboutWordServiceImpl(
     companion object {
         val log = LoggerFactory.getLogger(this::class.java)
     }
-}
-
-@ConfigurationProperties(prefix = "wordnik.api")
-data class WordinkKeyProperty @ConstructorBinding constructor(val key: String)
-
-@Component
-class ExternalWordSource(val webClient: WebClient, val wordnikKeyProperty: WordinkKeyProperty) {
-
-    final inline fun <reified T : GettingPartOfAdditionalInfoAboutWord> customGetInfoAboutWord(
-        wordSpelling: String,
-        link: String
-    ): String =
-        webClient.get()
-            .uri(
-                link, wordSpelling, wordnikKeyProperty.key
-            )
-            .retrieve()
-            .onStatus({ responseStatus ->
-                responseStatus == HttpStatus.NOT_FOUND
-            }) { throw ResponseStatusException(HttpStatus.NOT_FOUND) }
-            .bodyToFlux(T::class.java)
-            .blockFirst()
-            ?.partOfAdditionalInfoAboutWord ?: "Missing"
 }
